Fix bug in CSV file generation
This commit fixes a bug in the CSV file generation when the test
rate mode is used to analyze a trace.
More specifically, if no speed could be extracted from a message,
test_rate_speed was kept to None, and update_msg_same_speed.append()
was not called all the times, due to a missing "else".
This was making one of the lists (the one for "Same_speed_as_previous")
shorter than the others, causing zip to stop over the shortest input
iterable, thus "cutting" the CSV file in the middle of a trace.